Migrate `Find` and associated logic to `find.go` for improved organization; no functional changes.
